Preparations for non-player entities!
Added some base classes to get started on our class hierachy. They are
empty for now, but there are comments inside describing what type of
stuff should be in them.
I was unable to compile, but there are no errors shown in my IDE so it's
probably just a compiler issue. Will look into that after some sleep,
but if yall have problems compiling as well then i probably messed
something up when editing code tonight. Do not hesitate to revert this
commit if that is the case.